Create Basketball.py
This game is exactly what it is named: basketball.  A modification for Bombsquad that adds basketball to the existing sports:  football and hockey. Complete with hoop, ball, and referee, this mod is the beta version of what will soon have foul calls and a three point line.  However, it is completely playable as of now. Have fun!