Create Basketball2.0
This updated version of basketball now includes fouls, foul shots, jump balls, and a light-up hoop.  It has also taken care of the glitches in the beta version. This game is ready for servers, but the updates are not finished.  Up next will be a three-point line.  Until then, have fun playing with this version!