Final Update to Basketball.py
This is the last in a series of updates to the game Basketball. In it includes anti-cheating measures (such as: goal-tending prevention, foul shot safety precautions and the like) to keep game-play smooth.  This final version of the game includes: a hoop, a basketball, two-pointers, out-of-bounds, possession, jump-balls, a referee, and in newer updates: a three-point line with three-point-shots, goal-tending prevention, and foul-outs to keep the game about skill, not aggression.  Hope everyone enjoys it! Have fun.