Update SimonSays.py
Simon say is a classic game where the players have to do things told by someone, but only if "Simon says."  In this game there are a variety of colored circles laid across the map.  The players are then told to run to the different circles, the top, bottom, left, or right sides of the map, but if they do when Simon doesn't say to then they lose.  It gets faster as you play and it gets increasingly harder to determine what you are and aren't supposed to do.  Overall, a fun mod for Eric Froemling's game: Bombsquad.