Add Volcano Powerup and snowyPowerup.py
Volcano will spew bombs from the position where used. The user will die, but an impact bomb will fall onto every player's head from the sky.  Take from it what you will.  The snowyPowerup.py file is NOT MINE.  It is from the Bombsquad Official Mod-Manager.  Go there for the official version.  (It does not include [nor will include] Volcano.)  

Disclaimer: for this powerup to work, the snowyPowerup file must be installed, as well as the buddy bunny and snowballz powerups from the manager to avoid all errors.